refacotr: 로비 관련 기능 훅 분리
